Ability findByName() method in facade and service
Null checks in AbilityFacade arguments

+++ service-layer/src/main/java/cz/muni/fi/pa165/hauntedhouses/facade/HouseFacadeImpl.java
@@ -103,6 +103,10 @@ public class HouseFacadeImpl implements HouseFacade {
 
     @Override
     public void purge(HouseDTO house) {
+        if (house == null) {
+            throw new IllegalArgumentException("House is null.");
+        }
+
         House houseEntity = beanMappingService.mapTo(house, House.class);
 
         houseService.purge(houseEntity);
